Finishin populating db fixtures
