+++ user_profile_module/constants.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+# Enums
+class GenderEnum(str, Enum):
+    female = "female"
+    male = "male"
+    other = "other"
